pyproject.toml
[tool.poetry]
name = "iak-reporting-tool"
version = "0.1.0"
description = "Automated reporting tools for IAK (Instandhouding Advisering Kunstwerken) project - generates PI reports, attention points, and risk assessments from DISK exports"
authors = ["Arcadis IAK Team <iak-team@arcadis.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/ArjanvanLaarArcadis/IAK-reporting-tool"
repository = "https://github.com/ArjanvanLaarArcadis/IAK-reporting-tool"
documentation = "https://github.com/ArjanvanLaarArcadis/IAK-reporting-tool/blob/main/README.md"
keywords = ["reporting", "excel", "automation", "infrastructure", "IAK", "DISK"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Engineering",
]
packages = [{include = "IAK_Report", from = "."}]

[tool.poetry.dependencies]
python = "^3.12"
docx2pdf = "^0.1.8"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
pillow = "^11.1.0"
python-docx = "^1.1.2"
pywin32 = {version = "^309", markers = "sys_platform == 'win32'"}
pyxlsb = "^1.0.10"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
iak-generate-pi = "IAK_Report.generate_pi_rapportage:main"
iak-generate-attention = "IAK_Report.generate_aandachtspunten_beheerder:main"
iak-generate-risks = "IAK_Report.generate_hoogste_risicos:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
